+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 #### Enhancements
 
+* Remove MD5 hashing of configure options, not avialbale in FIPS mode. (#78)
 * Add experimental support for cmake-based projects.
 * Retry on HTTP failures during downloads. [#63] (Thanks, @jtarchie and @jvshahid!)
 * Support Ruby 2.4 frozen string literals.